fix(indexer-alt): stop early termination of watermark task
## Description

Checking whether the watermark task's receiver channel is closed is not
sufficient -- we also need to check whether it is empty, because all its
senders may be gone, but it needs to flush its queue as well.

## Test plan

This issue was easiest to reproduce by running a single pipeline that
ended with a large number of empty checkpoints, eg. `ev_struct_inst` up
to checkpoint 5000. Without this change, it would drop the last 100 or
so (empty) checkpoints, and with it, it will reliably update the
watermark to 5000 before exiting:

```
haneul$ cargo run -p haneul-indexer-alt --release --                                   \
  --database-url "postgres://postgres:postgrespw@localhost:5432/haneul_indexer_alt" \
  --remote-store-url https://checkpoints.mainnet.haneul.io                          \
  --last-checkpoint 5000 --pipeline ev_struct_inst
```