indexer-alt: introduce a "Pending" type
## Description

Split apart `Indexed<H>` into two parts -- `Indexed<H>` and
`Pending<H>`. Originally, the hope was that both the sequential and
concurrent pipelines could use the previous `Indexed<H>` and `collector`
implementation, but in hindsight, that was not a good idea, because the
existing collector is now designed to deal with chunks of data that are
not checkpoint-aligned, and sequential pipelines must write data out
checkpoint-by-checkpoint.

This change corrects that by splitting out the part of `Indexed<H>` that
dealt with splitting apart a single checkpoint's data into a dedicated
`Pending<H>` type that is internal to the concurrent collector.

## Test plan

Run the unit tests and run the pipeline on the first 5000 checkpoints:

```
haneul$ cargo nextest run -p haneul-indexer-alt
haneul$ cargo run -p haneul-indexer-alt --release --                                   \
  --database-url "postgres://postgres:postgrespw@localhost:5432/haneul_indexer_alt" \
  --remote-store-url https://checkpoints.mainnet.haneul.io                          \
  --last-checkpoint 5000
```