fix(indexer-alt): watermark update off-by-one
## Description

Fix a quirk of the watermark update logic in the concurrent pipeline,
where we would never issue an update for just the genesis checkpoint.

This fix was identified while working on the sequential pipeline where
we watermark updates and row updates are coupled and we therefore cannot
afford to ignore the first checkpoint (doing so would cause it to drop
the all the rows from the first checkpoint as well).

## Test plan

Run the indexer on just one checkpoint:

```
haneul$ cargo run -p haneul-indexer-alt --release --                                   \
  --database-url "postgres://postgres:postgrespw@localhost:5432/haneul_indexer_alt" \
  --remote-store-url https://checkpoints.mainnet.haneul.io                          \
  --last-checkpoint 1
```

It will write a watermark.