[move-ide] Avoid analyzing unmodified dependencies
## Description 

This PR implements an optimization to the symbolication algorithm that
avoids analyzing dependencies if they have not changed. Previously we
were avoiding re-compiling unchanged dependencies but we were still
analyzing them which introduces unnecessary overhead.

The implementation involved separating analysis of the main program and
the dependencies, and merging the results of the two together in the end
(whether the dependencies are computed fresh or obtained from the
cache). In particular, we now create two analysis visitors per analysis
phase, one for the main program and one for the dependencies.

For a simple package with Haneul framework as a dependency we observe a
significant reduction in analysis time.

Before:

![image](https://github.com/user-attachments/assets/06148508-9b22-4f52-b8c9-21968dcc4b1f)

After:

![image](https://github.com/user-attachments/assets/d14f6b5e-f2fa-4312-a1c2-81f68ce210c2)


## Test plan 

All existing test must pass. I also tested manually to verify that
reporting references (now merged between the main program and the
dependencies) works correctly.