[move] Add support for raw `abort`s in clever errors
## Description 

This adds support for raw `abort`s in a similar manner to raw
`assert!`ions. It will introduce the line number that the `abort` was
called on (and behaves in the same manner as `assert!` with no error
codes for macro line number remapping).

## Test plan 

Added parser tests as well as additional transactional tests, and
verified that they produce the exact same output as the derived line
number assertion tests that they copy.