indexer-alt: sum_packages pipeline
## Description

Pipeline for writing out the set of latest packages. This is used to do
type resolution during reads, and also to answer queries about package
versions.

Originally, this was called `kv_packages`, but I realised that because
of system packages, it is not append-only, so I renamed it to match the
other summary tables.

## Test plan

Ran the indexer with the new pipeline on the first 1.2M checkpoints.
This includes multiple system package upgrades and the first user
package publish.

Inspect the contents of the table at that point:

```
haneul$ cargo run -p haneul-indexer-alt --release --                                   \
  --database-url "postgres://postgres:postgrespw@localhost:5432/haneul_indexer_alt" \
  indexer --remote-store-url https://checkpoints.mainnet.haneul.io                  \
  --last-checkpoint 1200000 --pipeline sum_packages
```

```
haneul_indexer_alt=# SELECT package_id, original_id, package_version FROM sum_packages;
                             package_id                             |                            original_id                             | package_version
--------------------------------------------------------------------+--------------------------------------------------------------------+-----------------
 \x0000000000000000000000000000000000000000000000000000000000000001 | \x0000000000000000000000000000000000000000000000000000000000000001 |               1
 \x0000000000000000000000000000000000000000000000000000000000000002 | \x0000000000000000000000000000000000000000000000000000000000000002 |               2
 \x0000000000000000000000000000000000000000000000000000000000000003 | \x0000000000000000000000000000000000000000000000000000000000000003 |               3
 \x000000000000000000000000000000000000000000000000000000000000dee9 | \x000000000000000000000000000000000000000000000000000000000000dee9 |               2
 \x39ac04c24dbedf422abb8d582973ee733dbbab07a597fc98300666abe7982034 | \x39ac04c24dbedf422abb8d582973ee733dbbab07a597fc98300666abe7982034 |               1
```