indexer-alt: wal_obj_types pipeline
## Description

Adds the concurrent pipeline that writes the write-ahead log for
`sum_obj_types`. It re-uses the `process` implementation from
`sum_obj_types` and then writes it into an append-only table.

Note that today:

- The pipelines are completely independent, which means the WAL pipeline
redoes the processing work of the summary pipeline (this is presumably
not an issue because the repeated work is not particularly heavy).
- This change does not include the pruning necessary to keep this
table's size in check (in practice it should only be a couple of gigs in
size). This will come in a follow-up PR.

## Test plan

Run the indexer and cross check the live-object set calculated from the
write-ahead log with the summary:

```
haneul$ cargo run -p haneul-indexer-alt --release --                                   \
  --database-url "postgres://postgres:postgrespw@localhost:5432/haneul_indexer_alt" \
  indexer --remote-store-url https://checkpoints.mainnet.haneul.io                  \
  --last-checkpoint 5000
```

```
haneul_indexer_alt=# SELECT COUNT(*) FROM sum_obj_types;
 count
-------
   592
(1 row)

                                                  ^
haneul_indexer_alt=# SELECT
    COUNT(*)
FROM (
    SELECT DISTINCT ON (object_id)
        *
    FROM
        wal_obj_types
    ORDER BY
        object_id,
        object_version DESC
) o
WHERE
    o.owner_kind IS NOT NULL;
 count
-------
   592
(1 row)
```

## Stack

-  
-  

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: