[move-ide] Fixed auto-completion on the dot character
## Description 

This PR fixes a certain problem with auto-completion on the `.`
character. Previously, once a part of the access chain after the `.` was
parsed, auto-completion information for the `.` character location was
no longer generated. This is particularly problematic if someone expects
auto-completion in the middle of the function code, for example:
```move
public struct SomeStruct has drop {
    some_field: u64,
}

public fun foo(_param: SomeStruct) {}

public fun bar(): SomeStruct {
    let val = SomeStruct{some_field: 42};
    val.
    SomeStruct {some_field: 42}
}
```
It is expected that auto-completion would work correctly after `val.`
but previously it did not because this expression was actually parsed as
`val.SomeStruct`. This PR fixes this:

![image](https://github.com/user-attachments/assets/d7cb51d0-8faf-44a3-af16-fdb571459194)

Previously, auto-completion for the `.` location would also not work if
what follows the `.` is a function call (whether the function being
called was resolved correctly or not). It now works correctly in both
cases:

![image](https://github.com/user-attachments/assets/11c5e07c-5f89-48ea-a3cd-a8b3829565c0)

![image](https://github.com/user-attachments/assets/57e7209c-1afe-4381-9a42-35279045fa45)



## Test plan 

All new and (updated) all tests must pass