[Consensus] soft bundle to be included in a block atomically
## Description 

Following the work of https://github.com/GeunhwaJeong/haneul/pull/19572 ,
this PR is ensuring that a soft bundle will atomically be included in a
consensus block. That:
1. Strengthens the guarantees around the soft bundle sequencing to allow
them processed in exact sequence and not break in multiple blocks.
2. allow us to simply retry transactions of GC'ed blocks as we just
retry all the transactions of softbundle

Kept the solution simple for now and if a soft bundle doesn't fit to a
block then it's cached and is added to the next block.

## Test plan 

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: