Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/create-dapp@0.3.36

### Patch Changes

-   Updated dependencies [4c1da69]
    -   haneullabs/dapp-kit@0.14.35

## haneullabs/dapp-kit@0.14.35

### Patch Changes

-   4c1da69: Mark vanilla-extract libraries as dev dependencies

## haneullabs/deepbook-v3@0.12.9

### Patch Changes

-   0224a16: Governance function update