Version Packages
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and publish to npm
yourself or [setup this action to publish
automatically](https://github.com/changesets/action#with-publishing). If
you're not ready to do a release yet, that's fine, whenever you add more
changesets to main, this PR will be updated.


# Releases
## haneullabs/create-dapp@0.3.37

### Patch Changes

-   Updated dependencies [1c2046a]
    -   haneullabs/dapp-kit@0.14.36

## haneullabs/dapp-kit@0.14.36

### Patch Changes

-   1c2046a: Revert dev dependency change from v0.14.35 due to breakage