indexer-alt: generalize GraphQLConfig
## Description

`GraphQLConfig` is a derive macro for deriving `Serialize`,
`Deserialize`, and `Debug`, and introducing `#[serde(default = ...)]`
annotations that rely on the fields of the type's `Default`
implementation.

This allows config types to rely on their `Default` implementation as a
single source of truth for default values, and is used through-out
`haneul-graphql-rpc` to define its config types.

This change generalizes it (renames it) so it can be used in the indexer
as well.

## Test plan

Build and test `haneul-graphql-rpc`, `haneul-mvr-graphql-rpc`:

```
haneul$ cargo nextest run -p haneul-graphql-rpc
haneul$ cargo nextest run -p haneul-graphql-e2e-tests
haneul$ cargo nextest run -p haneul-mvr-graphql-rpc
```

This is a behaviour preserving change for the moment.