indexer-alt: enable pipelines by file-based config
## Description

Rather than deciding which pipelines to run according to a `--pipeline`
flag, decide them based on which pipelines have been configured in the
TOML file. It suffices that the pipeline is mentioned in the config (no
config fields need to be overridden for it).

Additionally, this change makes it so that the write-ahead log portion
of consistent tables is only written to when a checkpoint lag has been
specified.

## Test plan

Run indexers with the following configs:

```
[ingestion]
remote-store-url = "https://checkpoints.mainnet.haneul.io"

[pipeline.kv-objects]
[pipeline.kv-transactions]
```

```
[ingestion]
remote-store-url = "https://checkpoints.mainnet.haneul.io"

[consistency]
consistent-range = 1000

[pipeline.sum-obj-types]
```

```
[ingestion]
remote-store-url = "https://checkpoints.mainnet.haneul.io"

[committer]
collect-interval-ms = 1000

[pipeline.tx-calls]

[pipeline.tx-affected-objects]
collect-interval-ms = 5000
```