indexer-alt: generate example config
## Description

Using the default impl for `IndexerConfig` to generate the output for
`generate-config` lead to sub-par results because most of its fields are
optional, and so will not show up.

This change introduces an explicit `example` function that is
responsible for generating a non-default output that does a better job
of documenting what fields and pipelines are available to configure.

## Test plan

Generate a config:

```
haneul$ cargo run -p haneul-indexer-alt -- generate-config

[ingestion]
checkpoint-buffer-size = 5000
ingest-concurrency = 200
retry-interval-ms = 200

[consistency]
consistent-pruning-interval-ms = 300000
pruner-delay-ms = 120000

[committer]
write-concurrency = 5
collect-interval-ms = 500
watermark-interval-ms = 500

[pruner]
interval-ms = 300000
delay-ms = 120000
retention = 4000000
max-chunk-size = 2000

[pipeline.sum_coin_balances]

[pipeline.wal_coin_balances]

[pipeline.sum_obj_types]

[pipeline.wal_obj_types]

[pipeline.sum_displays]

[pipeline.sum_packages]

[pipeline.ev_emit_mod]

[pipeline.ev_struct_inst]

[pipeline.kv_checkpoints]

[pipeline.kv_epoch_ends]

[pipeline.kv_epoch_starts]

[pipeline.kv_feature_flags]

[pipeline.kv_objects]

[pipeline.kv_protocol_configs]

[pipeline.kv_transactions]

[pipeline.obj_versions]

[pipeline.tx_affected_addresses]

[pipeline.tx_affected_objects]

[pipeline.tx_balance_changes]

[pipeline.tx_calls]

[pipeline.tx_digests]

[pipeline.tx_kinds]

```