fix(haneul): be consistent about haneul config directory
## Description
When a `--config` is supplied and it is a directory, we treat it as the
haneul config directory in some but not all cases. This changes makes it so
that if `--config` does not look like a YAML file, then it is treated
like it should be the config directory, and that location is used
everywhere the config directory is expected.

## Test plan

Manually tested running `haneul start`, `haneul start --network.config ...`:

```sh
$ haneul start
$ haneul start --force-regenesis
$ haneul start --network.config ~/.haneul/haneul_config
$ haneul start --network.config ~/.haneul/haneul_config/network.yaml
$ haneul start --network.config "$CUSTOM_CONFIG"
$ haneul start --network.config "$CUSTOM_CONFIG/network.yaml"
```

A custom config directory can be generated by taking a copy of the
current config directory and tweaking it:

```sh
$ cp -r ~/.haneul/haneul_config $CUSTOM_CONFIG
$ rm "$CUSTOM_CONFIG/haneul.keystore" "$CUSTOM_CONFIG/haneul.aliases"
$ sed -i '' "s:$HOME/.haneul/haneul_config:$CUSTOM_CONFIG:" "$CUSTOM_CONFIG/network.yaml"
```

You can test that the local network is up and running by querying for
its chain identifier:

```sh
$ curl -LX POST  "http://localhost:9000" \
    --header 'Content-Type: application/json' \
    --data '{
  "jsonrpc": "2.0",
  "id": 1,
  "method": "haneul_getChainIdentifier",
  "params": []
}' | jq .
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Support fully overriding the config directory during `haneul
start`.
- [ ] Rust SDK:
- [ ] REST API: