refactor(indexer-alt): factor out Db/DbArgs
## Description

Pull out the `Db` and `DbArgs` types from `haneul-indexer-alt-framework`,
so that they can also be used in the reader implementation. They have
been combined with the existing `TempDb`, and its crate has been renamed
from `haneul-temp-pg-db` to `haneul-pg-db` to account for the fact that it's
now holding something more general.

## Test plan
CI