[move-compiler] Support for skipping function bodies during compilation
## Description 

This PR adds the ability for the compiler to skip compiling function
bodies for some files. This is particularly useful in the IDE setting
where we can cache results of compilation of the entire package and
re-compile only files changed by the user.

The difference for larger codebases, for example Haneul's Deepbook package,
is significant and directly affects user experience. Currently, it takes
over 1s to compile Deepbook package itself (dependencies are already
cached), which means that the user has to wait over 1s for
auto-completion request to provide completion hints:

![image](https://github.com/user-attachments/assets/c7de8950-ea11-48b1-b722-e3fce99bb5d0)

With this compiler change (and additional changes to `move-analyzer`
which will be landed in a separate PR), we get Deepbok's compilation
time down to less than 200ms:


![image](https://github.com/user-attachments/assets/e9f47e45-544d-4c54-9716-afd1ef85319c)


## Test plan 

All existing tests must pass