easy(indexer-alt): small clean-ups
## Description

- Row chunk size should default to `i16::MAX` if the field count is `0`.
- Fold the `if empty { ... }` check into the `debug_assert` so "nothing
gets left behind" when we're building for release (in reality the
compiler will probably not have a problem with compiling this out, but
it's good practice to fold the full condition into assert body).

## Test plan

CI

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:
- [ ] REST API: